Some final reorganizing of ITs, they're now a bit easier to run due to the auth/no-auth requirement being segregated into separate test classes. Finally closes #6
